+++ Assets/Model/character/NavMesh_AI.Animation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5e943670ca122c749a673553da82810c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
