+++ Assets/Model/character/NavMesh_AI.Interaction.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 68150edd2a58e3848ae460893b9a8bcd
+guid: 1e0530da8a295c94598cfef09eca2d5a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
