+++ Assets/Model/character/NavMesh_AI.Movement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f12ce3cb3639a06498b92e45bc65aef5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
