+++ Assets/Model/character/NavMesh_AI.StateManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 072f3292c1adab4408c5d46b57694a55
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
