+++ Assets/Model/character/NavMesh_AI.Utilities.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+public partial class NavMesh_AI
+{
+    internal void SetCooldown()
+    {
+        cooldownTimer = collisionCooldown;
+    }
+}
