+++ Assets/Model/character/NavMesh_AI.Utilities.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 00b2d08c05ac1e74bab41137c15e9c94
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
