+++ Assets/Model/character/NavMesh_AI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 85a6887f3b4feff41997343d5c4f725f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
